+++ docs/pages/api-docs/unstable-trap-focus.js
@@ -2,8 +2,8 @@ import React from 'react';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
 import { prepareMarkdown } from 'docs/src/modules/utils/parseMarkdown';
 
-const pageFilename = 'api/trap-focus';
-const requireRaw = require.context('!raw-loader!./', false, /\/trap-focus\.md$/);
+const pageFilename = 'api/unstable-trap-focus';
+const requireRaw = require.context('!raw-loader!./', false, /\/unstable-trap-focus\.md$/);
 
 export default function Page({ docs }) {
   return <MarkdownDocs docs={docs} />;
