+++ docs/pages/api-docs/unstable-trap-focus.md
@@ -1,0 +1,39 @@
+---
+filename: /packages/material-ui/src/Unstable_TrapFocus/Unstable_TrapFocus.js
+---
+
+<!--- This documentation is automatically generated, do not try to edit it. -->
+
+# Unstable_TrapFocus API
+
+<p class="description">The API documentation of the Unstable_TrapFocus React component. Learn more about the props and the CSS customization points.</p>
+
+## Import
+
+```js
+import Unstable_TrapFocus from '@material-ui/core/Unstable_TrapFocus';
+// or
+import { Unstable_TrapFocus } from '@material-ui/core';
+```
+
+You can learn more about the difference by [reading this guide](/guides/minimizing-bundle-size/).
+
+Utility component that locks focus inside the component.
+
+
+
+## Props
+
+| Name | Type | Default | Description |
+|:-----|:-----|:--------|:------------|
+| <span class="prop-name required">children&nbsp;*</span> | <span class="prop-type">element</span> |  |  |
+| <span class="prop-name">disableAutoFocus</span> | <span class="prop-type">bool</span> | <span class="prop-default">false</span> |  |
+| <span class="prop-name">disableEnforceFocus</span> | <span class="prop-type">bool</span> | <span class="prop-default">false</span> |  |
+| <span class="prop-name">disableRestoreFocus</span> | <span class="prop-type">bool</span> | <span class="prop-default">false</span> |  |
+| <span class="prop-name required">getDoc&nbsp;*</span> | <span class="prop-type">func</span> |  |  |
+| <span class="prop-name required">isEnabled&nbsp;*</span> | <span class="prop-type">func</span> |  |  |
+| <span class="prop-name required">open&nbsp;*</span> | <span class="prop-type">bool</span> |  |  |
+
+The component cannot hold a ref.
+
+
