+++ packages/material-ui/src/Unstable_TrapFocus/Unstable_TrapFocus.js
@@ -151,44 +151,12 @@ function TrapFocus(props) {
 }
 
 TrapFocus.propTypes = {
-  /**
-   * A single child content element.
-   */
   children: PropTypes.element.isRequired,
-  /**
-   * If `true`, the modal will not automatically shift focus to itself when it opens, and
-   * replace it to the last focused element when it closes.
-   * This also works correctly with any modal children that have the `disableAutoFocus` prop.
-   *
-   * Generally this should never be set to `true` as it makes the modal less
-   * accessible to assistive technologies, like screen readers.
-   */
   disableAutoFocus: PropTypes.bool,
-  /**
-   * If `true`, the modal will not prevent focus from leaving the modal while open.
-   *
-   * Generally this should never be set to `true` as it makes the modal less
-   * accessible to assistive technologies, like screen readers.
-   */
   disableEnforceFocus: PropTypes.bool,
-  /**
-   * If `true`, the modal will not restore focus to previously focused element once
-   * modal is hidden.
-   */
   disableRestoreFocus: PropTypes.bool,
-  /**
-   * Return the document to consider.
-   * We use it to implement the restore focus between different browser documents.
-   */
   getDoc: PropTypes.func.isRequired,
-  /**
-   * Do we still want to enforce the focus?
-   * This prop helps nesting TrapFocus elements.
-   */
   isEnabled: PropTypes.func.isRequired,
-  /**
-   * If `true`, the modal is open.
-   */
   open: PropTypes.bool.isRequired,
 };
