+++ packages/material-ui/src/Unstable_TrapFocus/Unstable_TrapFocus.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useFakeTimers } from 'sinon';
 import { expect } from 'chai';
-import TrapFocus from './TrapFocus';
+import TrapFocus from './Unstable_TrapFocus';
 import consoleErrorMock from 'test/utils/consoleErrorMock';
 import { createClientRender, fireEvent, screen } from 'test/utils/createClientRender';
